+++ market.engine/Transaction.cs
@@ -3,7 +3,7 @@
     public class Transaction
     {
         public SecurityNames Security { get; set; }
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public bool OnMargin { get; set; }
     }
 }
