+++ market.engine/tests/MarketPriceCard.test.cs
@@ -10,7 +10,7 @@ namespace market.engine.tests
     {
         public static void BullPriceTest()
         {
-            var expected = new int[] { 0, 18, 23, 11, 12, 46, 18, -5, 34, 15 };
+            var expected = new long[] { 0, 18, 23, 11, 12, 46, 18, -5, 34, 15 };
             var prices = MarketPriceCard.Generate(Situation.Bull, rvalue: 4);
 
             foreach(var security in Security.EnumerateAll())
@@ -21,7 +21,7 @@ namespace market.engine.tests
 
         public static void BearPriceTest()
         {
-            var expected = new int[] { 0, -8, -10, -10, -15, -20, -20, -23, -25, -7 };
+            var expected = new long[] { 0, -8, -10, -10, -15, -20, -20, -23, -25, -7 };
             var prices = MarketPriceCard.Generate(Situation.Bear, rvalue: 12);
 
             foreach (var security in Security.EnumerateAll())
